Made logfile, directory and app name dynamic
Log file, directory and application name can now be set from the
app.config (or web.config).  Use keyname 'LogFile' for the log file and
'Application' for the application name. If these parameters are not
found, the log service will use the application directory as the
location for the logfile and 'Unspecified application' as application
name.